requirements.txt
asgiref==3.3.1
boto3==1.16.25
botocore==1.19.25
dj-database-url==0.5.0
Django==3.1.3
django-appconf==1.0.4
django-crispy-forms==1.10.0
django-heroku==0.3.1
django-imagekit==4.0.2
django-storages==1.10.1
gunicorn==20.0.4
jmespath==0.10.0
pilkit==2.0
Pillow==8.0.1
psycopg2==2.8.6
python-dateutil==2.8.1
pytz==2020.4
s3transfer==0.3.3
six==1.15.0
sqlparse==0.4.1
urllib3==1.26.2
whitenoise==5.2.0
